policy setting ui moify
