+++ src/frontend/src/index.html
@@ -12,7 +12,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
-<body>
+<body style="background-color: #1B2A32;">
   <app-root></app-root>
 </body>
 </html>
